♻ :: 패키지 구조 변경 및 인조키 변경

+++ src/main/java/com/example/moizaspringserver/domain/user/repository/UserIntroduceLinkRepository.java
@@ -3,5 +3,5 @@ package com.example.moizaspringserver.domain.user.repository;
 import com.example.moizaspringserver.domain.user.entity.UserIntroduceLink;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserIntroduceLinkRepository extends CrudRepository<UserIntroduceLink, Long>{
+public interface UserIntroduceLinkRepository extends CrudRepository<UserIntroduceLink, Integer>{
 }
